+++ streams/src/main/java/org/apache/kafka/streams/kstream/internals/KStreamSlidingWindowAggregate.java
@@ -17,25 +17,17 @@
 package org.apache.kafka.streams.kstream.internals;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.metrics.Sensor;
-import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.streams.KeyValue;
-import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.kstream.Aggregator;
 import org.apache.kafka.streams.kstream.EmitStrategy;
-import org.apache.kafka.streams.kstream.EmitStrategy.StrategyType;
 import org.apache.kafka.streams.kstream.Initializer;
 import org.apache.kafka.streams.kstream.Window;
 import org.apache.kafka.streams.kstream.Windowed;
 import org.apache.kafka.streams.kstream.SlidingWindows;
-import org.apache.kafka.streams.kstream.internals.KStreamImplJoin.TimeTracker;
-import org.apache.kafka.streams.processor.api.ContextualProcessor;
 import org.apache.kafka.streams.processor.api.Processor;
 import org.apache.kafka.streams.processor.api.ProcessorContext;
 import org.apache.kafka.streams.processor.api.Record;
 import org.apache.kafka.streams.processor.api.RecordMetadata;
-import org.apache.kafka.streams.processor.internals.InternalProcessorContext;
-import org.apache.kafka.streams.processor.internals.metrics.StreamsMetricsImpl;
 import org.apache.kafka.streams.state.KeyValueIterator;
 import org.apache.kafka.streams.state.TimestampedWindowStore;
 import org.apache.kafka.streams.state.ValueAndTimestamp;
@@ -44,12 +36,6 @@ import org.slf4j.LoggerFactory;
 import java.util.HashSet;
 import java.util.Set;
 
-import static org.apache.kafka.streams.StreamsConfig.InternalConfig.EMIT_INTERVAL_MS_KSTREAMS_WINDOWED_AGGREGATION;
-import static org.apache.kafka.streams.processor.internals.metrics.ProcessorNodeMetrics.emitFinalLatencySensor;
-import static org.apache.kafka.streams.processor.internals.metrics.ProcessorNodeMetrics.emittedRecordsSensor;
-import static org.apache.kafka.streams.processor.internals.metrics.StreamsMetricsImpl.PROCESSOR_NODE_ID_TAG;
-import static org.apache.kafka.streams.processor.internals.metrics.StreamsMetricsImpl.maybeMeasureLatency;
-import static org.apache.kafka.streams.processor.internals.metrics.TaskMetrics.droppedRecordsSensor;
 import static org.apache.kafka.streams.state.ValueAndTimestamp.getValueOrNull;
 
 public class KStreamSlidingWindowAggregate<KIn, VIn, VAgg> implements KStreamAggProcessorSupplier<KIn, VIn, Windowed<KIn>, VAgg> {
@@ -85,7 +71,7 @@ public class KStreamSlidingWindowAggregate<KIn, VIn, VAgg> implements KStreamAgg
 
     @Override
     public Processor<KIn, VIn, Windowed<KIn>, Change<VAgg>> get() {
-        return new KStreamSlidingWindowAggregateProcessor();
+        return new KStreamSlidingWindowAggregateProcessor(storeName, emitStrategy, sendOldValues);
     }
 
     public SlidingWindows windows() {
@@ -97,60 +83,14 @@ public class KStreamSlidingWindowAggregate<KIn, VIn, VAgg> implements KStreamAgg
         sendOldValues = true;
     }
 
-    private class KStreamSlidingWindowAggregateProcessor extends ContextualProcessor<KIn, VIn, Windowed<KIn>, Change<VAgg>> {
-        private TimestampedWindowStore<KIn, VAgg> windowStore;
-        private TimestampedTupleForwarder<Windowed<KIn>, VAgg> tupleForwarder;
-        private Sensor droppedRecordsSensor;
-        private Sensor emittedRecordsSensor;
-        private Sensor emitFinalLatencySensor;
-        private long observedStreamTime = ConsumerRecord.NO_TIMESTAMP;
-        private long lastEmitCloseTime = ConsumerRecord.NO_TIMESTAMP;
+    private class KStreamSlidingWindowAggregateProcessor extends AbstractKStreamTimeWindowAggregateProcessor<KIn, VIn, VAgg> {
+
         private Boolean reverseIteratorPossible = null;
-        private InternalProcessorContext<Windowed<KIn>, Change<VAgg>> internalProcessorContext;
-        private final TimeTracker timeTracker = new TimeTracker();
-        private final Time time = Time.SYSTEM;
+        private long observedStreamTime = ConsumerRecord.NO_TIMESTAMP;
 
-        @Override
-        public void init(final ProcessorContext<Windowed<KIn>, Change<VAgg>> context) {
-            super.init(context);
-            internalProcessorContext = (InternalProcessorContext<Windowed<KIn>, Change<VAgg>>) context;
-            final StreamsMetricsImpl metrics = internalProcessorContext.metrics();
-            final String threadId = Thread.currentThread().getName();
-            droppedRecordsSensor = droppedRecordsSensor(threadId, context.taskId().toString(), metrics);
-            emittedRecordsSensor = emitFinalLatencySensor(threadId, context.taskId().toString(),
-                internalProcessorContext.currentNode().name(), metrics);
-            emitFinalLatencySensor = emitFinalLatencySensor(threadId, context.taskId().toString(),
-                internalProcessorContext.currentNode().name(), metrics);
-            windowStore = context.getStateStore(storeName);
-            if (emitStrategy.type() == StrategyType.ON_WINDOW_CLOSE) {
-                // Don't set flush lister which emit cache results
-                tupleForwarder = new TimestampedTupleForwarder<>(
-                    windowStore,
-                    context,
-                    sendOldValues);
-            } else {
-                tupleForwarder = new TimestampedTupleForwarder<>(
-                    windowStore,
-                    context,
-                    new TimestampedCacheFlushListener<>(context),
-                    sendOldValues);
-            }
-
-            log.info("EmitStrategy=" + emitStrategy.type());
-            // Restore last emit close time for ON_WINDOW_CLOSE strategy
-            if (emitStrategy.type() == StrategyType.ON_WINDOW_CLOSE) {
-                final Long lastEmitTime = internalProcessorContext.processorMetadataForKey(storeName);
-                if (lastEmitTime != null) {
-                    lastEmitCloseTime = lastEmitTime;
-                }
-                final long emitInterval = StreamsConfig.InternalConfig.getLong(
-                    context.appConfigs(),
-                    EMIT_INTERVAL_MS_KSTREAMS_WINDOWED_AGGREGATION,
-                    1000L
-                );
-                timeTracker.setEmitInterval(emitInterval);
-                log.info("EmitInterval=" + emitInterval);
-            }
+        protected KStreamSlidingWindowAggregateProcessor(final String storeName,
+            final EmitStrategy emitStrategy, final boolean sendOldValues) {
+            super(storeName, emitStrategy, sendOldValues);
         }
 
         @Override
@@ -232,7 +172,7 @@ public class KStreamSlidingWindowAggregate<KIn, VIn, VAgg> implements KStreamAgg
                 processInOrder(record, closeTime);
             }
 
-            maybeMeasureLatency(() -> maybeForwardFinalResult(record, closeTime), time, emitFinalLatencySensor);
+            maybeMeasureEmitFinalLatency(record, closeTime);
         }
 
         public void processInOrder(final Record<KIn, VIn> record, final long closeTime) {
@@ -519,70 +459,24 @@ public class KStreamSlidingWindowAggregate<KIn, VIn, VAgg> implements KStreamAgg
             return rightWinAgg != null && rightWinAgg.timestamp() > inputRecordTimestamp;
         }
 
-        private void maybeForwardUpdate(final Record<KIn, VIn> record,
-                                  final Window window,
-                                  final VAgg oldAgg,
-                                  final VAgg newAgg,
-                                  final long newTimestamp) {
-
-            if (emitStrategy.type() == StrategyType.ON_WINDOW_CLOSE) {
-                return;
-            }
-            tupleForwarder.maybeForward(
-                record.withKey(new Windowed<>(record.key(), window))
-                    .withValue(new Change<>(newAgg, sendOldValues ? oldAgg : null))
-                    .withTimestamp(newTimestamp));
-        }
-
-        private void maybeForwardFinalResult(final Record<KIn, VIn> record, final long closeTime) {
-            if (emitStrategy.type() != StrategyType.ON_WINDOW_CLOSE) {
+        @Override
+        protected void maybeForwardFinalResult(final Record<KIn, VIn> record, final long closeTime) {
+            if (!shouldEmitFinal(closeTime)) {
                 return;
             }
 
-            final long now = internalProcessorContext.currentSystemTimeMs();
-            // Throttle emit frequency
-            if (now < timeTracker.nextTimeToEmit) {
-                return;
-            }
-
-            // Schedule next emit time based on now to avoid the case that if system time jumps a lot,
-            // this can be triggered everytime
-            timeTracker.nextTimeToEmit = now;
-            timeTracker.advanceNextTimeToEmit();
-
-            // Close time does not progress
-            if (lastEmitCloseTime != ConsumerRecord.NO_TIMESTAMP && lastEmitCloseTime >= closeTime) {
-                return;
-            }
-
-            final long emitWindowStart = closeTime - windows.timeDifferenceMs();
+            final long emitRangeUpperBoundExclusive = closeTime - windows.timeDifferenceMs();
             // When closeTime is on right boundary of a window, the window is not closed
-            // So if emitWindowStart is 0, we shouldn't emit
-            if (emitWindowStart <= 0) {
+            // So if emitRangeUpperBoundExclusive is 0, we shouldn't emit
+            if (emitRangeUpperBoundExclusive <= 0) {
                 return;
             }
 
-            final long lastEmitWindowStart = lastEmitCloseTime == ConsumerRecord.NO_TIMESTAMP ?
+            final long emitRangeLowerBoundInclusive = lastEmitCloseTime == ConsumerRecord.NO_TIMESTAMP ?
                 0L : lastEmitCloseTime - windows.timeDifferenceMs();
 
             // Fetch to emitWindowStart - 1 so avoid close right boundary of window
-            final KeyValueIterator<Windowed<KIn>, ValueAndTimestamp<VAgg>> windowToEmit =  windowStore
-                .fetchAll(lastEmitWindowStart, emitWindowStart - 1);
-
-            int emittedCount = 0;
-            while (windowToEmit.hasNext()) {
-                emittedCount++;
-                final KeyValue<Windowed<KIn>, ValueAndTimestamp<VAgg>> kv = windowToEmit.next();
-                tupleForwarder.maybeForward(
-                    record.withKey(kv.key)
-                        .withValue(new Change<>(kv.value.value(), null))
-                        .withTimestamp(kv.value.timestamp())
-                        .withHeaders(null)); // Don't set header
-            }
-            emittedRecordsSensor.record(emittedCount);
-
-            lastEmitCloseTime = closeTime;
-            internalProcessorContext.addProcessorMetadataKeyValue(storeName, closeTime);
+            fetchAndEmit(record, closeTime, emitRangeLowerBoundInclusive, emitRangeUpperBoundExclusive - 1);
         }
 
         private void updateWindowAndForward(final Window window,
